Tooling/hotwire_turbo (#19)
* Remove Autogenerated Files

Remove files that were generated by the webpack boilerplate and are
not needed in our application.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>

* Rename Entry Point to turbo_drive

We rename the entry point for the frontend from "app" to "turbo_drive" to
better reflect its purpose. This includes renaming the JS and CSS
files as well.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>

* Install Hotwired Turbo

npm install @hotwired/turbo

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>

* Add Turbo To the frontend

Import Turbo and create an event listener for turbo:load.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>

---------

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>